header added to Makefile, strlcat return type fixed, calloc changed to return same as malloc

+++ ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: rpohlen <rpohlen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/09 22:35:01 by rpohlen           #+#    #+#             */
-/*   Updated: 2021/11/24 19:37:01 by rpohlen          ###   ########.fr       */
+/*   Updated: 2021/11/25 16:01:27 by rpohlen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcat(char *dest, const char *src, size_t n)
+size_t	ft_strlcat(char *dest, const char *src, size_t n)
 {
 	size_t	dstlen;
 	size_t	srclen;
